@@ -1,3 +1,3 @@
 module mit2
 
-go 1.15
+go 1.19
