@@ -10,7 +10,12 @@ package main
 // Please do not change this file.
 //
 
-import "mit2/src/mr"
+import (
+	"mit2/src/mr"
+	"strconv"
+	"strings"
+	"unicode"
+)
 import "plugin"
 import "os"
 import "fmt"
@@ -22,7 +27,26 @@ func main() {
 		os.Exit(1)
 	}
 
-	mapf, reducef := loadPlugin(os.Args[1]) // os.Args 保存了调用命令的参数，第[0] 是 "mrworker.go";第[1]个是"wc.so"
+	mapf := func(filename string, contents string) []mr.KeyValue {
+		// function to detect word separators.
+		ff := func(r rune) bool { return !unicode.IsLetter(r) }
+
+		// split contents into an array of words.
+		words := strings.FieldsFunc(contents, ff)
+
+		kva := []mr.KeyValue{}
+		for _, w := range words {
+			kv := mr.KeyValue{w, "1"}
+			kva = append(kva, kv)
+		}
+		return kva
+	}
+
+	reducef := func(key string, values []string) string {
+		// return the number of occurrences of this word.
+		return strconv.Itoa(len(values))
+	}
+	//mapf, reducef := loadPlugin(os.Args[1]) // os.Args 保存了调用命令的参数，第[0] 是 "mrworker.go";第[1]个是"wc.so"
 
 	mr.Worker(mapf, reducef)
 }
@@ -31,11 +55,13 @@ func main() {
 // load the application Map and Reduce functions
 // from a plugin file, e.g. ../mrapps/wc.so
 //
+
 func loadPlugin(filename string) (func(string, string) []mr.KeyValue, func(string, []string) string) {
 	p, err := plugin.Open(filename)
 	if err != nil {
-		log.Fatalf("cannot load plugin %v", filename)
+		log.Printf("cannot load plugin %v,error message：\n%v", filename, err)
 	}
+	log.Println("插件的地址为：", p)
 	xmapf, err := p.Lookup("Map")
 	if err != nil {
 		log.Fatalf("cannot find Map in %v", filename)
